refactor: extract main into bin
