+++ crates/typos-lsp/Cargo.toml
@@ -11,7 +11,10 @@ edition.workspace = true
 rust-version.workspace = true
 include.workspace = true
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[[bin]]
+name = "typos-lsp"
+path = "src/bin/typos-lsp/main.rs"
+doc = false
 
 [dependencies]
 anyhow = "1.0.70"
